+++ deca/builder.py
@@ -133,6 +133,7 @@ class Builder:
             vfs: VfsDatabase,
             vpath_complete_map,
             symlink_changed_file,
+            do_not_build_archive,
     ):
         print(f'build_node: {dst_path} | {src_path} | {vnode.file_type} | {vnode.v_path}')
 
@@ -144,7 +145,8 @@ class Builder:
             _, src_file = os.path.split(src_path)
 
         if vnode.file_type == FTYPE_SARC:
-            self.build_node_sarc(dst_path, src_path, vnode, vfs, vpath_complete_map, symlink_changed_file)
+            if not do_not_build_archive:
+                self.build_node_sarc(dst_path, src_path, vnode, vfs, vpath_complete_map, symlink_changed_file)
         elif src_path is None:
             pass  # no source path,
         elif src_file.find('DECA') >= 0:
@@ -178,6 +180,7 @@ class Builder:
             dst_path: str,
             subset=None,
             symlink_changed_file=False,
+            do_not_build_archive=False,
     ):
         print(f'build_node: {dst_path} | {src_path}')
 
@@ -310,6 +313,7 @@ class Builder:
                             vfs=vfs,
                             vpath_complete_map=vpaths_completed,
                             symlink_changed_file=symlink_changed_file,
+                            do_not_build_archive=do_not_build_archive,
                         )
 
             if not any_change and len(depends) > 0: