+++ deca/gui/build.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+pyside2-uic main_window.ui > main_window.py